requirements.txt
Django==1.9.4
django-annoying==0.8.7
Pillow==3.1.1
